models/test_models/boxes/requirements.txt
pillow
python-socketio[client]
